@@ -18,16 +18,16 @@
     "url": "https://github.com/awslabs/aws-crt-nodejs/issues"
   },
   "devDependencies": {
-    "@types/jquery": "^3.3.30",
+    "@types/jquery": "^3.3.31",
     "source-map-loader": "^0.2.4",
-    "ts-loader": "^6.0.4",
-    "typescript": "^3.5.3",
-    "webpack": "^4.35.3",
-    "webpack-cli": "^3.3.6"
+    "ts-loader": "^6.2.1",
+    "typescript": "^3.7.3",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10"
   },
   "dependencies": {
     "aws-crt": "file:../../../",
-    "aws-sdk": "^2.497.0",
+    "aws-sdk": "^2.585.0",
     "jquery": "^3.4.1"
   }
 }
