@@ -12,16 +12,27 @@
  * express or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-const os = require('os');
-const path = require('path');
-const fs = require('fs');
+import * as path from 'path';
+import { platform, arch } from 'os';
+import { existsSync } from 'fs';
+import { versions } from 'process';
+
+const upgrade_string = "Please upgrade to node >=10.13.0, or use the provided browser implementation.";
+if ('napi' in versions) {
+    const napi_version = parseInt(versions['napi']);
+    if (napi_version < 4) {
+        throw new Error("The AWS CRT native implementation requires that NAPI version 4 be present. " + upgrade_string);
+    }
+} else {
+    throw new Error("The current runtime is not reporting an NAPI version. " + upgrade_string);
+}
 
 const binary_name = 'aws-crt-nodejs';
-const platformDir = `${os.platform}-${os.arch}`;
+const platformDir = `${platform}-${arch}`;
 
 let source_root = path.resolve(__dirname, '..', '..');
 const dist = path.join(source_root, 'dist');
-if (fs.existsSync(dist)) {
+if (existsSync(dist)) {
     source_root = dist;
 }
 
@@ -34,7 +45,7 @@ const search_paths = [
 
 let binding;
 for (const path of search_paths) {
-    if (fs.existsSync(path + '.node')) {
+    if (existsSync(path + '.node')) {
         binding = require(path);
         break;
     }
